+++ ReadingOutputFromFLYonPIC_v2/__init__.py
@@ -24,10 +24,17 @@ from . import ChargeStateColors
 # scan with SCFLY
 from . import ScanSCFLY
 
+# processing routines
+from . import LoadFLYonPICData
+from . import LoadSCFLYData
+from . import ReduceToPerChargeState
+
 #plotting libraries
 from . import PlotAtomicPopulations
 from . import PlotTimingData
 from . import PlotSummarySCFLYScan
+
+
 
 __all__ = [
     "Config",
@@ -35,7 +42,9 @@ __all__ = [
     "Reader",
     "ChargeStateColors",
     "ScanSCFLY",
-    "ScanInitialState",
+    "LoadFLYonPICData",
+    "LoadSCFLYData",
+    "ReduceToPerChargeState",
     "PlotAtomicPopulations",
     "PlotTimingData",
     "PlotSummarySCFLYScan"]
